Wheee...lotts of changes. Stage2 can now boot boot images created with mkuuubimage, and will make a simple one for testing with the "boot" command.
Also, the bootsector has been updated to support pseudo-binary indirection. The
max filesize it can handle is b^2 / 4 + 4b where b = block size, in bytes. For
512B blocks used on floppies, that's 67.584 bytes.

Lastly, the default now is the textual console. The graphical one is still in
there, but some changes to the bootsector would be required to make it work
again. The new boot model makes booting much simpler, and vga is just such a
pain it's not worth the effort. Booting will be so incredibally fast that snap
will be loaded in under 2s anyway, so shouldn't be a problem :)