added an "mbinfo" command to print information passed from a multiboot bootloader, and made a bunch of various tweaks to accomadate it.
The command buffer is now placed above 1M and is explicitly reserved in a .bss
to prevent multiboot information being placed in the same place. By default,
the buffer is limited to 512 chars.

Also, the relocation of the boot code is delayed until the last possible time
since grub likes to put all the multiboot information below 1M.